split up the testing
